requirements.txt
appdirs==1.4.3
asn1crypto==1.1.0
astroid==2.3.1
attrs==19.2.0
black==19.3b0
cffi==1.12.3
Click==7.0
colorama==0.4.1
cryptography==2.7
Django==2.2.6
django-rest-knox==4.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
psycopg2==2.8.3
pycparser==2.19
pylint==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2